Made include/ctf.hpp define CTF_VERSION 1.22
Changed the copy constructor for tensors, as there was a conflict for a symmetry modifying constructor. To get rid of the default copy constructor it was necessary to define both templated and parent (untemplated) class constructors rather than just the latter.

This will break codes that used Tensor B(A, false/true), which now needs to be Tensor B(false/true, A)