Add support for Hadamard indices in contractions with sparse tensors. However, current method is not performant, does e.g. for
c[i] = a[i]*b[i]

compare size of a and b (based on max of nonzero and min of predicted number of rows for CSR for sparse and size for dense), and if a is smaller do

create sparse matrix A

A[i,i] = a[i]

c[i] = A[i,j]*b[j]

many indices handled by doing this transformation recursively. In the future, should support these contractions by doing batching.

closes #23 and closes #47